+++ crates/blockifier/src/blockifier/stateful_validator.rs
@@ -17,7 +17,7 @@ use crate::state::errors::StateError;
 use crate::state::state_api::StateReader;
 use crate::transaction::account_transaction::AccountTransaction;
 use crate::transaction::errors::{TransactionExecutionError, TransactionPreValidationError};
-use crate::transaction::transaction_execution::Transaction;
+use crate::transaction::transactions::Transaction;
 use crate::transaction::transactions::ValidatableTransaction;
 
 #[cfg(test)]
