+++ crates/blockifier/src/transaction.rs
@@ -6,7 +6,6 @@ pub mod errors;
 pub mod objects;
 #[cfg(any(feature = "testing", test))]
 pub mod test_utils;
-pub mod transaction_execution;
 pub mod transaction_types;
 pub mod transaction_utils;
 pub mod transactions;
